# PERF: infer_datetime_format without padding #11142