Closes #11142