@chris-b1 2015-09-19
@sinhrks pushed changes for your notes.  The perf impact on `_guess_datetime_format` is negligible because it is only ever called once.  Nothing showed up in `asv`, here are a couple timings.

    In [1]: s, l = pd.date_range('2014-1-1', periods=10), pd.date_range('2014-1-1', periods=100000)
    In [2]: s = s.strftime('%Y-%m-%d')
    In [3]: l = l.strftime('%Y-%m-%d')

    # master
    In [4]: %timeit pd.to_datetime(l, infer_datetime_format=True)
    10 loops, best of 3: 30.3 ms per loop

    In [5]: %timeit pd.to_datetime(s, infer_datetime_format=True)
    1000 loops, best of 3: 300 µs per loop

    # PR
    In [5]: %timeit pd.to_datetime(l, infer_datetime_format=True)
    10 loops, best of 3: 30.3 ms per loop

    In [6]: %timeit pd.to_datetime(s, infer_datetime_format=True)
    1000 loops, best of 3: 308 µs per loop